# Conflicts:
#	src/main/webapp/view/viewBlog.jsp

+++ src/main/java/com/team09/controller/admin/DeleteBlogServlet.java
@@ -17,8 +17,16 @@ public class DeleteBlogServlet extends HttpServlet {
         BlogService blogService = BlogServiceImpl.getInstance();
         String blogId = request.getParameter("blogId");
 
-        Blog blog = blogService.getBlogById(blogId);
-        blogService.deleteBlogs(blog);
+        if (blogId == null || blogId.isEmpty()){
+            return;
+        }else {
+            Blog blog = blogService.getBlogById(blogId);
+            blogService.deleteBlogs(blog);
+        }
+
+        String page = request.getParameter("page");
+        response.sendRedirect(request.getContextPath() + "/" + page);
+
     }
 
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
